🏗♻️ Bento Auto-Envelope (#34820)
Enables Bento components to run on documents without `v0.js`

It specifies a `bento` extension wrapper for immediate execution, absent the AMP runtime. In this case, we use barebones implementations of `BaseElement` and `CustomElement` that are bundled as part of the extension. 

To support this mounting scheme, we make two additional infra changes:

1. Lint rule `restrict-this-access`, so that Preact-based components only use methods available in the barebones `BaseElement`.

2. Update `babel-plugin-transform-amp-extension-call` to unwrap `AMP.extension()`, to avoid providing it as part of the Bento wrapper.

For [browser support](https://caniuse.com/?search=custom%20element), this change also provides a `custom-elements-polyfill.js` that can be used once per document.