🏗️ 🔥  Fix batchedReads being shared by different callers (#36978)
* Fix batchedReads being shared by different callers

In the old code, `batchedRead` performed a file read, and hashed the contents with an input `optionsHash`. But `batchedRead` is designed to allow multiple callers to share the same file, and they may provide different `optionsHash`es. When this happened, only the first caller's read+`optionsHash` would be used, and the 2nd+ caller would receive the first caller's hashes!

This simple fix is to move this out of `batchedRead` and let that just _read_ and hash the file contents. If you want to rehash with options, do that afterwards.

* Fix incorrect reference to file hash

* Fix files with duplicated contents compiling incorrectly