Extract AMPPreactBaseElement (#37309)
* extract fn to register custom element

* create (empty) amp pbe

* inherit amp-accordion from amp-pbe

* fix custom-element definitions for iframes (fixes tests)

* workarounds to properly test bento components

we don't want AMP.BaseElement to be declared so that it won't get injected into the class hierarchy. current tests assume that AMP should be initialized all the time which breaks bento tests

* split out generic bento testing from amp tests

current (hacky) workarounds: some tests depend on CSS, which we expect consumers to install. This differs from the AMP components that register styles themselves. To workaround this we have to import and append the bento component styles. This problem might be unique to components that don't use shadowCss. For components that do use shadowDom and have critical styling that needs to be tested, those styles are already written in JSS and declared as shadowCss in the bento component's declaration, which ensures that that the styles are baked into the component.

* Revert "workarounds to properly test bento components"

un-revert in order to test bento component

* inherit all non-video bento-amp components from amp-pbe

* rename VideoBaseElement -> AmpVideoBaseElement

* refactor inheritance hierarchy for video components to remove amp code from bento hierarchy

* fix amp-facebook hierarchy to preserve AmpFbBase functionality

* update docs to demonstrate new way to scaffold amp-* components

* move (some) amp-specific methods into amp-pbe

* update dependencies, rename BentoVideoBaseElement

* add doc

* refactor dailymotion

* ampvideobaseelement docs

* lint

* fix typings and lint

* remove file

* remove dict() based on ongoign discussions to deprecate

* disable local/restrict-this-access for the whole file

* remove more references to dict()