🏗 Apply Babel config to React Storybook (#37457)
Run the same React-specific transforms on Storybook files as we do on component bundles.

In effect, this maps Preact-style prop names to React-style. This prevents issues like ``Invalid DOM property `class`. Did you mean `className`?``

Note that we still get an error for using `tabindex` instead of `tabIndex`. This is because our prop mapping is erroneously flipped. See #37459