🏗️  Fix the `sources` of extension sourcemaps (#37829)
* Fix the `sources` of extension sourcemaps

We're still getting incorrect RTVs in our sourcemaps, which I thought I had fixed. The issue before was that the `sources` paths contained a `../src/...`, where `..` is relative to the map file. We fixed this by chopping off the `../`.

But this only worked for `v0.js` and other entrypoint files. For extension maps, the map file is written to `dist/v0/FOO.js.map`, and the `sources` paths start with `../../src/...`. This fixes all maps by resolving the relative path into a `cwd` relative source, which can then be joined with the `sourceRoot` without issue.

* lint

* Fix stdio